dagster/requirements.txt
dagster==1.5.0
dagit==1.5.0
requests==2.31.0
yfinance==0.2.40
psycopg2-binary==2.9.6
pymongo==4.3.3
milvus==2.2.0
transformers==4.35.0
torch==2.0.1
loguru==0.6.0
python-dotenv==1.0.0
aiohttp==3.8.5
phidata==2.0.0
stable-baselines3==2.0.0
tensorboard==2.14.0
minio==7.1.14
etcd3==0.12.0